+++ Projects/STM32WB5MM-DK/Applications/Zigbee/Zigbee_TempMeas_Server_Coord/Core/Src/app_stts22h.c
@@ -137,7 +137,7 @@ void STTS22H_PrintValues(void){
 
   STTS22H_TEMP_GetTemperature(&stts22h_obj_0,&temp_val);
 
-  sprintf(text,"T:%.1f C",temp_val);
+  sprintf(text,"T:%.1f C",(double)temp_val);
   UTIL_LCD_DisplayStringAt(0, LINE(3), (uint8_t *)text, CENTER_MODE);
   BSP_LCD_Refresh(0);
